tests: Create dedicated shader helper file
