+++ tests/framework/pipeline_helper.h
@@ -14,6 +14,7 @@
 
 #include "layer_validation_tests.h"
 #include "descriptor_helper.h"
+#include "shader_helper.h"
 
 // Helper class for tersely creating create pipeline tests
 //
