+++ tests/framework/render.cpp
@@ -1467,104 +1467,3 @@ std::array<std::array<uint32_t, 16>, 16> VkImageObj::Read() {
     stagingImage.UnmapMemory();
     return m;
 }
-
-VkPipelineShaderStageCreateInfo const &VkShaderObj::GetStageCreateInfo() const { return m_stage_info; }
-
-VkShaderObj::VkShaderObj(VkRenderFramework *framework, const char *source, VkShaderStageFlagBits stage, const spv_target_env env,
-                         SpvSourceType source_type, const VkSpecializationInfo *spec_info, char const *entry_point, bool debug, const void *pNext)
-    : m_framework(*framework), m_device(*(framework->DeviceObj())), m_source(source), m_spv_env(env) {
-    m_stage_info = vku::InitStructHelper();
-    m_stage_info.flags = 0;
-    m_stage_info.stage = stage;
-    m_stage_info.module = VK_NULL_HANDLE;
-    m_stage_info.pName = entry_point;
-    m_stage_info.pSpecializationInfo = spec_info;
-    if (source_type == SPV_SOURCE_GLSL) {
-        InitFromGLSL(debug, pNext);
-    } else if (source_type == SPV_SOURCE_ASM) {
-        InitFromASM();
-    }
-}
-
-bool VkShaderObj::InitFromGLSL(bool debug, const void *pNext) {
-    std::vector<uint32_t> spv;
-    m_framework.GLSLtoSPV(&m_device.phy().limits_, m_stage_info.stage, m_source, spv, debug, m_spv_env);
-
-    VkShaderModuleCreateInfo moduleCreateInfo = vku::InitStructHelper();
-    moduleCreateInfo.pNext = pNext;
-    moduleCreateInfo.codeSize = spv.size() * sizeof(uint32_t);
-    moduleCreateInfo.pCode = spv.data();
-
-    init(m_device, moduleCreateInfo);
-    m_stage_info.module = handle();
-    return VK_NULL_HANDLE != handle();
-}
-
-// Because shaders are currently validated at pipeline creation time, there are test cases that might fail shader module
-// creation due to supplying an invalid/unknown SPIR-V capability/operation. This is called after VkShaderObj creation when
-// tests are found to crash on a CI device
-VkResult VkShaderObj::InitFromGLSLTry(bool debug, const vkt::Device *custom_device) {
-    std::vector<uint32_t> spv;
-    // 99% of tests just use the framework's VkDevice, but this allows for tests to use custom device object
-    // Can't set at contructor time since all reference members need to be initialized then.
-    VkPhysicalDeviceLimits limits = (custom_device) ? custom_device->phy().limits_ : m_device.phy().limits_;
-    m_framework.GLSLtoSPV(&limits, m_stage_info.stage, m_source, spv, debug, m_spv_env);
-
-    VkShaderModuleCreateInfo moduleCreateInfo = vku::InitStructHelper();
-    moduleCreateInfo.codeSize = spv.size() * sizeof(uint32_t);
-    moduleCreateInfo.pCode = spv.data();
-
-    const auto result = init_try(((custom_device) ? *custom_device : m_device), moduleCreateInfo);
-    m_stage_info.module = handle();
-    return result;
-}
-
-bool VkShaderObj::InitFromASM() {
-    vector<uint32_t> spv;
-    m_framework.ASMtoSPV(m_spv_env, 0, m_source, spv);
-
-    VkShaderModuleCreateInfo moduleCreateInfo = vku::InitStructHelper();
-    moduleCreateInfo.codeSize = spv.size() * sizeof(uint32_t);
-    moduleCreateInfo.pCode = spv.data();
-
-    init(m_device, moduleCreateInfo);
-    m_stage_info.module = handle();
-    return VK_NULL_HANDLE != handle();
-}
-
-VkResult VkShaderObj::InitFromASMTry() {
-    vector<uint32_t> spv;
-    m_framework.ASMtoSPV(m_spv_env, 0, m_source, spv);
-
-    VkShaderModuleCreateInfo moduleCreateInfo = vku::InitStructHelper();
-    moduleCreateInfo.codeSize = spv.size() * sizeof(uint32_t);
-    moduleCreateInfo.pCode = spv.data();
-
-    const auto result = init_try(m_device, moduleCreateInfo);
-    m_stage_info.module = handle();
-    return result;
-}
-
-// static
-std::unique_ptr<VkShaderObj> VkShaderObj::CreateFromGLSL(VkRenderFramework *framework, const char *source,
-                                                         VkShaderStageFlagBits stage, const spv_target_env spv_env,
-                                                         const VkSpecializationInfo *spec_info, const char *entry_point,
-                                                         bool debug) {
-    auto shader =
-        std::make_unique<VkShaderObj>(framework, source, stage, spv_env, SPV_SOURCE_GLSL_TRY, spec_info, entry_point, debug);
-    if (VK_SUCCESS == shader->InitFromGLSLTry(debug)) {
-        return shader;
-    }
-    return {};
-}
-
-// static
-std::unique_ptr<VkShaderObj> VkShaderObj::CreateFromASM(VkRenderFramework *framework, const char *source,
-                                                        VkShaderStageFlagBits stage, const spv_target_env spv_env,
-                                                        const VkSpecializationInfo *spec_info, const char *entry_point) {
-    auto shader = std::make_unique<VkShaderObj>(framework, source, stage, spv_env, SPV_SOURCE_ASM_TRY, spec_info, entry_point);
-    if (VK_SUCCESS == shader->InitFromASMTry()) {
-        return shader;
-    }
-    return {};
-}
