+++ tests/framework/render.h
@@ -351,45 +351,3 @@ class VkImageObj : public vkt::Image {
     uint32_t m_mipLevels;
     uint32_t m_arrayLayers;
 };
-
-// What is the incoming source to be turned into VkShaderModuleCreateInfo::pCode
-typedef enum {
-    SPV_SOURCE_GLSL,
-    SPV_SOURCE_ASM,
-    // TRY == Won't try in contructor as need to be called as function that can return the VkResult
-    SPV_SOURCE_GLSL_TRY,
-    SPV_SOURCE_ASM_TRY,
-} SpvSourceType;
-
-class VkShaderObj : public vkt::ShaderModule {
-  public:
-    // optional arguments listed order of most likely to be changed manually by a test
-    VkShaderObj(VkRenderFramework *framework, const char *source, VkShaderStageFlagBits stage,
-                const spv_target_env env = SPV_ENV_VULKAN_1_0, SpvSourceType source_type = SPV_SOURCE_GLSL,
-                const VkSpecializationInfo *spec_info = nullptr, char const *entry_point = "main", bool debug = false,
-                const void *pNext = nullptr);
-    VkPipelineShaderStageCreateInfo const &GetStageCreateInfo() const;
-
-    bool InitFromGLSL(bool debug = false, const void *pNext = nullptr);
-    VkResult InitFromGLSLTry(bool debug = false, const vkt::Device *custom_device = nullptr);
-    bool InitFromASM();
-    VkResult InitFromASMTry();
-
-    // These functions return a pointer to a newly created _and initialized_ VkShaderObj if initialization was successful.
-    // Otherwise, {} is returned.
-    static std::unique_ptr<VkShaderObj> CreateFromGLSL(VkRenderFramework *framework, const char *source,
-                                                       VkShaderStageFlagBits stage, const spv_target_env = SPV_ENV_VULKAN_1_0,
-                                                       const VkSpecializationInfo *spec_info = nullptr,
-                                                       const char *entry_point = "main", bool debug = false);
-    static std::unique_ptr<VkShaderObj> CreateFromASM(VkRenderFramework *framework, const char *source, VkShaderStageFlagBits stage,
-                                                      const spv_target_env spv_env = SPV_ENV_VULKAN_1_0,
-                                                      const VkSpecializationInfo *spec_info = nullptr,
-                                                      const char *entry_point = "main");
-
-  protected:
-    VkPipelineShaderStageCreateInfo m_stage_info;
-    VkRenderFramework &m_framework;
-    vkt::Device &m_device;
-    const char *m_source;
-    spv_target_env m_spv_env;
-};
