+++ tests/unit/ray_tracing.cpp
@@ -16,6 +16,7 @@
 #include "../framework/ray_tracing_objects.h"
 #include "../framework/ray_tracing_nv.h"
 #include "../framework/descriptor_helper.h"
+#include "../framework/shader_helper.h"
 #include "../layers/utils/vk_layer_utils.h"
 
 void RayTracingTest::OOBRayTracingShadersTestBody(bool gpu_assisted) {
