+++ tests/unit/ray_tracing_gpu.cpp
@@ -15,6 +15,7 @@
 #include "../framework/ray_tracing_objects.h"
 #include "../framework/ray_tracing_nv.h"
 #include "../framework/descriptor_helper.h"
+#include "../framework/shader_helper.h"
 
 TEST_F(NegativeGpuAssistedRayTracing, ArrayOOBRayTracingShaders) {
     TEST_DESCRIPTION(
